Descriptions
In this case, an extra memory term d is added to the update rule for the mid-prices pi(t) of the dealers in the stochastic dealer model. The constant d determines whether the dealers are trend-followers (d>0) or contrarians (d<0). Under these settings, I programmed this code to simulate the 2D random walk in the stochastic dealer model with a memory term.